# api/LiteWeightApiDocumentation/public/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "LiteWeight API",
    "version": "v1"
  },
  "paths": {
...
    "/exercises": {
      "post": {
        "tags": [
          "Exercises"
        ],
        "summary": "Create Exercise",
        "description": "Creates an exercise to be owned by the authenticated user. The name of the exercise must not already exist for the user.<br/><br/><div style=\"background-color: #2f3031;padding: 12px 0px 12px 12px;border-left: 5px solid #F90258;\">Potential 400 Errors: <b>AlreadyExists, InvalidRequest, MaxLimit</b></div>",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetExerciseRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/SetExerciseRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/SetExerciseRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OwnedExerciseResponse"
                }
              }
            }
          },
          "401": {
            "description": "Authentication Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnauthorizedResponse"
                }
              }
            }
          },
          "403": {
            "description": "Authorization Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ForbiddenResponse"
                }
              }
            }
          },
          "429": {
            "description": "Too many requests",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TooManyRequestsResponse"
                }
              }
            }
          },
          "500": {
            "description": "Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServerErrorResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BadRequestResponse"
                }
              }
            }
          }
        }
      }
    },
    "/exercises/{exerciseId}": {
      "put": {
        "tags": [
          "Exercises"
        ],
        "summary": "Update Exercise",
        "description": "Updates an exercise owned by the authenticated user, if it exists. Note that any new name for the exercise must not already exist.<br/><br/><div style=\"background-color: #2f3031;padding: 12px 0px 12px 12px;border-left: 5px solid #F90258;\">Potential 400 Errors: <b>AlreadyExists, InvalidRequest</b></div>",
        "parameters": [
          {
            "name": "exerciseId",
            "in": "path",
            "description": "Id of the exercise to update",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "description": "Request",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetExerciseRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/SetExerciseRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/SetExerciseRequest"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResourceNotFoundResponse"
                }
              }
            }
          },
          "401": {
            "description": "Authentication Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnauthorizedResponse"
                }
              }
            }
          },
          "403": {
            "description": "Authorization Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ForbiddenResponse"
                }
              }
            }
          },
          "429": {
            "description": "Too many requests",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TooManyRequestsResponse"
                }
              }
            }
          },
          "500": {
            "description": "Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServerErrorResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BadRequestResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Exercises"
        ],
        "summary": "Delete Exercise",
        "description": "Deletes an exercise owned by the authenticated user. Removes the deleted exercise from any workout it was a part of.",
        "parameters": [
          {
            "name": "exerciseId",
            "in": "path",
            "description": "Id of the exercise to delete",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResourceNotFoundResponse"
                }
              }
            }
          },
          "401": {
            "description": "Authentication Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnauthorizedResponse"
                }
              }
            }
          },
          "403": {
            "description": "Authorization Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ForbiddenResponse"
                }
              }
            }
          },
          "429": {
            "description": "Too many requests",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TooManyRequestsResponse"
                }
              }
            }
          },
          "500": {
            "description": "Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServerErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BadRequestResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "description": "Message describing the error.",
            "nullable": true,
            "example": "An example error message"
          },
          "errorType": {
            "type": "string",
            "description": "Specific type of error.",
            "nullable": true,
            "example": "InvalidRequest"
          },
          "requestErrors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ModelBindingError"
            },
            "description": "Lists any errors associated when binding the request.",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "ForbiddenResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "description": "Response message.",
            "nullable": true,
            "example": "Forbidden access"
          }
        },
        "additionalProperties": false
      },
...
      "ModelBindingError": {
        "type": "object",
        "properties": {
          "property": {
            "type": "string",
            "description": "Property that is causing the error.",
            "nullable": true,
            "example": "name"
          },
          "message": {
            "type": "string",
            "description": "Error message for problematic property.",
            "nullable": true,
            "example": "The name property cannot be empty"
          }
        },
        "additionalProperties": false
      },
      "OwnedExerciseResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Id of the exercise.",
            "nullable": true,
            "example": "88a54457-2253-404e-ac09-82a8f2ce5fb8"
          },
          "name": {
            "type": "string",
            "description": "Name of the exercise.",
            "nullable": true,
            "example": "Barbell Curl"
          },
          "defaultWeight": {
            "type": "number",
            "description": "Default weight of the exercise (lb). Value that the exercise will be defaulted to when adding it to a workout.",
            "format": "double",
            "example": 65
          },
          "defaultSets": {
            "type": "integer",
            "description": "Default sets of the exercise. Value that the exercise will be defaulted to when adding it to a workout.",
            "format": "int32",
            "example": 3
          },
          "defaultReps": {
            "type": "integer",
            "description": "Default reps of the exercise. Value that the exercise will be defaulted to when adding it to a workout.",
            "format": "int32",
            "example": 15
          },
          "defaultDetails": {
            "type": "string",
            "description": "Default details of the exercise. Value that the exercise will be defaulted to when adding it to a workout.",
            "nullable": true,
            "example": "Don't overextend arms."
          },
          "videoUrl": {
            "type": "string",
            "description": "Video url of the exercise. Suggested use case is a video of how to perform the exercise.",
            "nullable": true,
            "example": "https://www.youtube.com/watch?v=kwG2ipFRgfo"
          },
          "focuses": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "List of focuses of the exercise.",
            "nullable": true,
            "example": [
              "Biceps",
              "Forearms"
            ]
          },
          "workouts": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OwnedExerciseWorkoutResponse"
            },
            "description": "List of workouts this exercise is associated with.",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "OwnedExerciseWorkoutResponse": {
        "type": "object",
        "properties": {
          "workoutId": {
            "type": "string",
            "description": "Id of the workout.",
            "nullable": true,
            "example": "b7c3c321-d5a0-4448-978c-0a8b7709ceb2"
          },
          "workoutName": {
            "type": "string",
            "description": "Name of the workout.",
            "nullable": true,
            "example": "Main Workout"
          }
        },
        "additionalProperties": false
      },
...
      "ResourceNotFoundResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "description": "Response message.",
            "nullable": true,
            "example": "The workout with id=cfbb1fd4-6d29-465c-9ea0-4938407ac65b was not found"
          }
        },
        "additionalProperties": false
      },
...
      "ServerErrorResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "description": "Response message.",
            "nullable": true,
            "example": "The server has encountered an error"
          }
        },
        "additionalProperties": false
      },
...
      "SetExerciseRequest": {
        "required": [
          "defaultReps",
          "defaultSets",
          "defaultWeight",
          "focuses",
          "name"
        ],
        "type": "object",
        "properties": {
          "name": {
            "maxLength": 40,
            "minLength": 1,
            "type": "string",
            "description": "Name of the exercise.",
            "example": "Bench Press"
          },
          "defaultWeight": {
            "maximum": 99999,
            "minimum": 0,
            "type": "number",
            "description": "Default weight of the exercise (lb). Value that the exercise will be defaulted to when adding it to a workout.",
            "format": "double",
            "example": 225.5
          },
          "defaultSets": {
            "maximum": 99,
            "minimum": 0,
            "type": "integer",
            "description": "Default sets of the exercise. Value that the exercise will be defaulted to when adding it to a workout.",
            "format": "int32",
            "example": 3
          },
          "defaultReps": {
            "maximum": 999,
            "minimum": 0,
            "type": "integer",
            "description": "Default reps of the exercise. Value that the exercise will be defaulted to when adding it to a workout.",
            "format": "int32",
            "example": 12
          },
          "focuses": {
            "maxItems": 20,
            "minItems": 1,
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "List of focuses of the exercise.",
            "example": [
              "Chest",
              "Strength Training"
            ]
          },
          "defaultDetails": {
            "maxLength": 120,
            "type": "string",
            "description": "Default details of the exercise. Value that the exercise will be defaulted to when adding it to a workout.",
            "nullable": true,
            "example": "Make sure to get a spotter."
          },
          "videoUrl": {
            "maxLength": 200,
            "type": "string",
            "description": "Video URL of the exercise. Suggested use case is a video of how to perform the exercise.",
            "nullable": true,
            "example": "https://www.youtube.com/watch?v=rT7DgCr-3pg"
          }
        },
        "additionalProperties": false
      },
...
      "TooManyRequestsResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "description": "Response message.",
            "nullable": true,
            "example": "Too many requests. Please try again later."
          }
        },
        "additionalProperties": false
      },
      "UnauthorizedResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "description": "Response message.",
            "nullable": true,
            "example": "Unauthorized access"
          }
        },
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "BearerDefinition": {
        "type": "http",
        "description": "Token authentication. \n\n 'Bearer TOKEN'\n\nBearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiaWF0IjoxNTE2MjM5MDIyfQ.SflKxwRJSMeKKF2QT4fwpMeJf36POk6yJV_adQssw5c",
        "scheme": "Bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}